Update fix.qc_status.by.source.R
Values of zero were only supposed to be dropped for experimental records/dose response summary values. UPDATE statement to set qc_status to 'fail' where toxval_numeric<=0 was edited to reflect the correct logic. This will prevent the error from happening to new records in the future, but does not add the missed MCLG=0 records back to the database. I don't believe logic is in place for a case where a record is failed errantly, so I believe this source will have to be reloaded at the least.